[fix]: refactor and attempt to fix caching logic
